reorg tool importing and ut
