+++ mmlmtools/tools/image_editing/__init__.py
@@ -3,8 +3,8 @@
 from .extension import ImageExtension
 from .remove import ObjectRemove
 from .replace import ObjectReplace
-from .stylization import InstructPix2Pix
+from .stylization import ImageStylization
 
 __all__ = [
-    'ImageExtension', 'ObjectRemove', 'ObjectReplace', 'InstructPix2Pix'
+    'ImageExtension', 'ObjectRemove', 'ObjectReplace', 'ImageStylization'
 ]
