+++ mmlmtools/tools/image_editing/stylization.py
@@ -39,7 +39,7 @@ def load_instruct_pix2pix(model, device):
     return instruct_pix2pix
 
 
-class InstructPix2Pix(BaseTool):
+class ImageStylization(BaseTool):
     DEFAULT_TOOLMETA = dict(
         name='Instruct Image Using Text',
         model={'model': 'timbrooks/instruct-pix2pix'},
