+++ mmlmtools/tools/parsers/__init__.py
@@ -1,11 +1,11 @@
 # Copyright (c) OpenMMLab. All rights reserved.
 from .base_parser import BaseParser
-from .custom_parsers import HuggingFaceAgentParser, LangchainParser
+from .custom_parsers import HuggingFaceAgentParser, LangChainParser
 from .naive_parser import NaiveParser
 from .type_mapping_parser import TypeMappingParser
 from .utils import Audio
 
 __all__ = [
-    'BaseParser', 'NaiveParser', 'TypeMappingParser', 'LangchainParser',
+    'BaseParser', 'NaiveParser', 'TypeMappingParser', 'LangChainParser',
     'HuggingFaceAgentParser', 'Audio'
 ]
