+++ tests/test_tools/test_image_canny/test_image_to_canny.py
@@ -1,30 +1,27 @@
 import os.path as osp
-from unittest import skipIf
 
 import cv2
 import numpy as np
-from mmengine import is_installed
 from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
-@skipIf(not is_installed('mmdet'), reason='requires mmdet')
-class TestTextToBbox(ToolTestCase):
+class TestImageToCanny(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'TextToBbox', parser=VisualChatGPTParser(), device='cuda')
+            'ImageToCanny', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
-        res = tool(f'{img_path}, man')
+        res = tool(img_path)
         assert isinstance(res, str)
 
         img = Image.fromarray(img)
         tool = load_tool(
-            'TextToBbox', parser=HuggingFaceAgentParser(), device='cuda')
-        res = tool(img, 'man')
+            'ImageToCanny', parser=HuggingFaceAgentParser(), device='cuda')
+        res = tool(img)
         assert isinstance(res, Image.Image)
