+++ tests/test_tools/test_image_depth/test_depth_to_image.py
@@ -6,22 +6,22 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
-class TestImageToScribble(ToolTestCase):
+class TestDepthTextToImage(ToolTestCase):
 
-    def test_call(self):
+    def test_all(self):
         tool = load_tool(
-            'ImageToScribble', parser=VisualChatGPTParser(), device='cuda')
+            'DepthTextToImage', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
-        res = tool(img_path)
+        res = tool(img_path, 'prompt')
         assert isinstance(res, str)
 
         img = Image.fromarray(img)
         tool = load_tool(
-            'ImageToScribble', parser=HuggingFaceAgentParser(), device='cuda')
-        res = tool(img)
+            'DepthTextToImage', parser=HuggingFaceAgentParser(), device='cuda')
+        res = tool(img, 'prompt')
         assert isinstance(res, Image.Image)
