+++ tests/test_tools/test_image_depth/test_image_to_depth.py
@@ -1,21 +1,19 @@
 import os.path as osp
-from unittest import skipIf
 
 import cv2
 import numpy as np
-from mmengine import is_installed
 from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
-@skipIf(not is_installed('mmpretrain'), reason='requires mmpretrain')
-class TestImageCaption(ToolTestCase):
+class TestImageToDepth(ToolTestCase):
 
     def test_call(self):
-        tool = load_tool('ImageCaption', device='cuda')
+        tool = load_tool(
+            'ImageToDepth', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
@@ -24,6 +22,6 @@ class TestImageCaption(ToolTestCase):
 
         img = Image.fromarray(img)
         tool = load_tool(
-            'ImageCaption', input_style='pil image', device='cuda')
+            'ImageToDepth', parser=HuggingFaceAgentParser(), device='cuda')
         res = tool(img)
-        assert isinstance(res, str)
+        assert isinstance(res, Image.Image)
