+++ tests/test_tools/test_image_editing/test_image_stylization.py
@@ -5,7 +5,7 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
 @skipIf(not is_installed('diffusers'), reason='requires diffusers')
@@ -13,13 +13,13 @@ class TestInstructPix2Pix(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'InstructPix2Pix', parser=VisualChatGPTParser(), device='cuda')
+            'ImageStylization', parser=LangChainParser(), device='cuda')
         img_path = 'tests/data/images/dog-image.jpg'
         res = tool(f'{img_path}, watercolor painting')
         assert isinstance(res, str)
 
         img = Image.open(img_path)
         tool = load_tool(
-            'InstructPix2Pix', parser=HuggingFaceAgentParser(), device='cpu')
+            'ImageStylization', parser=HuggingFaceAgentParser(), device='cpu')
         res = tool(img, 'watercolor painting')
         assert isinstance(res, Image.Image)
