+++ tests/test_tools/test_image_editing/test_object_replace.py
@@ -5,7 +5,7 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
 @skipIf(
@@ -15,7 +15,7 @@ class TestObjectReplace(ToolTestCase):
     def test_call(self):
         img_path = 'tests/data/images/test.jpg'
         tool = load_tool(
-            'ObjectReplace', parser=VisualChatGPTParser(), device='cpu')
+            'ObjectReplace', parser=LangChainParser(), device='cpu')
         res = tool(img_path, 'dog', 'a cartoon dog')
         assert isinstance(res, str)
