+++ tests/test_tools/test_image_pose/test_facelandmark.py
@@ -8,7 +8,7 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
 @skipIf(not is_installed('mmpose'), reason='requires mmpose')
@@ -16,7 +16,7 @@ class TestHumanFaceLandmark(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'HumanFaceLandmark', parser=VisualChatGPTParser(), device='cuda')
+            'HumanFaceLandmark', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
