+++ tests/test_tools/test_image_pose/test_image_to_pose.py
@@ -1,19 +1,22 @@
 import os.path as osp
+from unittest import skipIf
 
 import cv2
 import numpy as np
+from mmengine import is_installed
 from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
-class TestImageToCanny(ToolTestCase):
+@skipIf(not is_installed('mmpose'), reason='requires mmpose')
+class TestHumanBodyPose(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'ImageToCanny', parser=VisualChatGPTParser(), device='cuda')
+            'HumanBodyPose', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
@@ -22,6 +25,6 @@ class TestImageToCanny(ToolTestCase):
 
         img = Image.fromarray(img)
         tool = load_tool(
-            'ImageToCanny', parser=HuggingFaceAgentParser(), device='cuda')
+            'HumanBodyPose', parser=HuggingFaceAgentParser(), device='cuda')
         res = tool(img)
         assert isinstance(res, Image.Image)
