+++ tests/test_tools/test_image_pose/test_pose_to_image.py
@@ -8,7 +8,7 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
 @skipIf(not is_installed('mmagic'), reason='requires mmagic')
@@ -16,7 +16,7 @@ class TestPoseToImage(ToolTestCase):
 
     def test_all(self):
         tool = load_tool(
-            'PoseToImage', parser=VisualChatGPTParser(), device='cuda')
+            'PoseToImage', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
