+++ tests/test_tools/test_image_scribble/test_image_to_scribble.py
@@ -6,24 +6,22 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
-class TestScribbleTextToImage(ToolTestCase):
+class TestImageToScribble(ToolTestCase):
 
-    def test_all(self):
+    def test_call(self):
         tool = load_tool(
-            'ScribbleTextToImage', parser=VisualChatGPTParser(), device='cuda')
+            'ImageToScribble', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
-        res = tool(img_path, 'prompt')
+        res = tool(img_path)
         assert isinstance(res, str)
 
         img = Image.fromarray(img)
         tool = load_tool(
-            'ScribbleTextToImage',
-            parser=HuggingFaceAgentParser(),
-            device='cuda')
-        res = tool(img, 'prompt')
+            'ImageToScribble', parser=HuggingFaceAgentParser(), device='cuda')
+        res = tool(img)
         assert isinstance(res, Image.Image)
