+++ tests/test_tools/test_image_scribble/test_scribble_to_image.py
@@ -1,32 +1,29 @@
 import os.path as osp
-from unittest import skipIf
 
 import cv2
 import numpy as np
-from mmengine import is_installed
 from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
-@skipIf(not is_installed('mmdet'), reason='requires mmdet')
-class TestObjectDetectionTool(ToolTestCase):
+class TestScribbleTextToImage(ToolTestCase):
 
-    def test_call(self):
+    def test_all(self):
         tool = load_tool(
-            'ObjectDetectionTool', parser=VisualChatGPTParser(), device='cuda')
+            'ScribbleTextToImage', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
-        res = tool(img_path)
+        res = tool(img_path, 'prompt')
         assert isinstance(res, str)
 
+        img = Image.fromarray(img)
         tool = load_tool(
-            'ObjectDetectionTool',
+            'ScribbleTextToImage',
             parser=HuggingFaceAgentParser(),
             device='cuda')
-        img = Image.fromarray(img)
-        res = tool(img)
+        res = tool(img, 'prompt')
         assert isinstance(res, Image.Image)
