+++ tests/test_tools/test_image_text/test_image_to_text.py
@@ -1,19 +1,20 @@
 import os.path as osp
+from unittest import skipIf
 
 import cv2
 import numpy as np
+from mmengine import is_installed
 from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
 
 
-class TestImageToDepth(ToolTestCase):
+@skipIf(not is_installed('mmpretrain'), reason='requires mmpretrain')
+class TestImageCaption(ToolTestCase):
 
     def test_call(self):
-        tool = load_tool(
-            'ImageToDepth', parser=VisualChatGPTParser(), device='cuda')
+        tool = load_tool('ImageCaption', device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
@@ -22,6 +23,6 @@ class TestImageToDepth(ToolTestCase):
 
         img = Image.fromarray(img)
         tool = load_tool(
-            'ImageToDepth', parser=HuggingFaceAgentParser(), device='cuda')
+            'ImageCaption', input_style='pil image', device='cuda')
         res = tool(img)
-        assert isinstance(res, Image.Image)
+        assert isinstance(res, str)
