+++ tests/test_tools/test_image_text/test_text_to_image.py
@@ -5,7 +5,7 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
 @skipIf(not is_installed('mmagic'), reason='requires mmagic')
@@ -13,7 +13,7 @@ class TestTextToImage(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'TextToImage', parser=VisualChatGPTParser(), device='cuda')
+            'TextToImage', parser=LangChainParser(), device='cuda')
         res = tool('generate an image of a cat')
         assert isinstance(res, str)
