+++ tests/test_tools/test_imagebind/test_anything_to_image.py
