+++ tests/test_tools/test_ocr/test_ocr.py
@@ -8,14 +8,14 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
 @skipIf(not is_installed('mmocr'), reason='requires mmocr')
 class TestOCR(ToolTestCase):
 
     def test_call(self):
-        tool = load_tool('OCR', parser=VisualChatGPTParser(), device='cuda')
+        tool = load_tool('OCR', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
@@ -33,7 +33,7 @@ class TestImageMaskOCR(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'ImageMaskOCR', parser=VisualChatGPTParser(), device='cuda')
+            'ImageMaskOCR', parser=LangChainParser(), device='cuda')
         res = tool('tests/data/images/test-image.png, '
                    'tests/data/images/test-mask.png')
         assert isinstance(res, str)
