+++ tests/test_tools/test_segmentation/test_segment_anything.py
@@ -4,7 +4,7 @@ from mmengine import is_installed
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import VisualChatGPTParser
+from mmlmtools.tools.parsers import LangChainParser
 
 
 @skipIf(
@@ -13,7 +13,7 @@ class TestSegmentAnything(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'SegmentAnything', parser=VisualChatGPTParser(), device='cpu')
+            'SegmentAnything', parser=LangChainParser(), device='cpu')
         res = tool('tests/data/images/test-image.png')
         assert isinstance(res, str)
 
@@ -24,7 +24,7 @@ class TestSegmentClicked(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'SegmentClicked', parser=VisualChatGPTParser(), device='cpu')
+            'SegmentClicked', parser=LangChainParser(), device='cpu')
         res = tool('tests/data/images/test-image.png, '
                    'tests/data/images/test-mask.png')
         assert isinstance(res, str)
@@ -36,6 +36,6 @@ class TestObjectSegmenting(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'ObjectSegmenting', parser=VisualChatGPTParser(), device='cpu')
+            'ObjectSegmenting', parser=LangChainParser(), device='cpu')
         res = tool('tests/data/images/test-image.png, water cup')
         assert isinstance(res, str)
