+++ tests/test_tools/test_vqa/test_vqa.py
@@ -8,15 +8,15 @@ from PIL import Image
 
 from mmlmtools import load_tool
 from mmlmtools.testing import ToolTestCase
-from mmlmtools.tools.parsers import HuggingFaceAgentParser, VisualChatGPTParser
+from mmlmtools.tools.parsers import HuggingFaceAgentParser, LangChainParser
 
 
-@skipIf(not is_installed('mmagic'), reason='requires mmagic')
-class TestCannyTextToImage(ToolTestCase):
+@skipIf(not is_installed('mmpretrain'), reason='requires mmpretrain')
+class TestVisionQuestionAnswering(ToolTestCase):
 
     def test_call(self):
         tool = load_tool(
-            'CannyTextToImage', parser=VisualChatGPTParser(), device='cuda')
+            'VisualQuestionAnswering', parser=LangChainParser(), device='cuda')
         img = np.ones([224, 224, 3]).astype(np.uint8)
         img_path = osp.join(self.tempdir.name, 'temp.jpg')
         cv2.imwrite(img_path, img)
@@ -25,6 +25,8 @@ class TestCannyTextToImage(ToolTestCase):
 
         img = Image.fromarray(img)
         tool = load_tool(
-            'CannyTextToImage', parser=HuggingFaceAgentParser(), device='cuda')
+            'VisualQuestionAnswering',
+            parser=HuggingFaceAgentParser(),
+            device='cuda')
         res = tool(img, 'prompt')
-        assert isinstance(res, Image.Image)
+        assert isinstance(res, str)
